+++ src/main/java/net/andrecarbajal/naviMusic/audio/GuildMusicManager.java
@@ -3,15 +3,16 @@ package net.andrecarbajal.naviMusic.audio;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import lombok.Getter;
+import net.dv8tion.jda.api.entities.Guild;
 
 @Getter
 public class GuildMusicManager {
     public final AudioPlayer player;
     public final TrackScheduler scheduler;
 
-    public GuildMusicManager(AudioPlayerManager manager) {
-        player=manager.createPlayer();
-        scheduler=new TrackScheduler(player);
+    public GuildMusicManager(AudioPlayerManager manager, Guild guild) {
+        player = manager.createPlayer();
+        scheduler = new TrackScheduler(player, guild);
         player.addListener(scheduler);
     }