+++ src/main/java/net/andrecarbajal/naviMusic/util/URLUtils.java
@@ -1,12 +1,15 @@
 package net.andrecarbajal.naviMusic.util;
 
+import lombok.experimental.UtilityClass;
+
 import java.net.URI;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+@UtilityClass
 public class URLUtils {
-    public static Optional<String> getURLParam(String url, String param) {
+    public Optional<String> getURLParam(String url, String param) {
         final String regex = param + "=([^&]*)";
 
         Pattern pattern = Pattern.compile(regex);
@@ -15,7 +18,7 @@ public class URLUtils {
         return matcher.find() ? Optional.of(matcher.group(1)) : Optional.empty();
     }
 
-    public static boolean isURL(String input) {
+    public boolean isURL(String input) {
         try {
             new URI(input);
             return true;